requirements.txt
aniso8601==7.0.0
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
coverage==6.3.2
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-oauth-toolkit==2.0.0
django-rest-auth==0.9.5
django-rest-authemail==2.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
fastapi==0.75.2
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
jwcrypto==1.3
oauthlib==3.2.0
Pillow==9.1.0
promise==2.3
pycparser==2.21
pydantic==1.9.0
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
text-unidecode==1.3
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
